NWJS Window support
if NWJS is enabled, get some info about the window, put a border around
it, and then add a workaround for some weird window sizing bugs. The
window is getting made 1 pixel wider than it's told to if it's
resizable, and way bigger if it isn't. This commit hacks around that by
setting window to be resizable, 799x599